Separate IOB fuzzer
